python-for-multimedia/extract-video-metadata/requirements.txt
ffmpeg-python